requirements.txt
pandas>=1.3.0
python-dotenv>=0.19.0
Pillow>=8.3.0
requests>=2.26.0
requests-oauthlib>=1.3.0
schedule>=1.1.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Platform-specific libraries (uncomment as needed)
# tweepy>=4.10.0  # Twitter/X
# facebook-sdk>=3.1.0  # Facebook
# instabot>=0.117.0  # Instagram
# linkedin-api>=2.0.0  # LinkedIn
# TikTokApi>=5.2.0  # TikTok (unofficial)

# Image generation
# openai>=0.27.0  # For DALL-E image generation
